+++ lib/majic/once.ex
@@ -1,13 +1,13 @@
-defmodule Majic.Helpers do
+defmodule Majic.Once do
   @moduledoc """
   Contains convenience functions for one-off use.
   """
 
-  alias Majic.Result
   alias Majic.Server
 
-  @spec perform_once(Path.t() | {:bytes, binary}, [Server.option()]) ::
-          {:ok, Result.t()} | {:error, term()}
+  @process_timeout Majic.Config.default_process_timeout()
+
+  @spec perform(Majic.target(), [Server.start_option()], timeout()) :: Majic.result()
 
   @doc """
   Runs a one-shot process without supervision.
@@ -16,13 +16,13 @@ defmodule Majic.Helpers do
 
   ## Example
 
-      iex(1)> {:ok, result} = Majic.Helpers.perform_once(".")
+      iex(1)> {:ok, result} = Majic.Once.perform(".")
       iex(2)> result
       %Majic.Result{content: "directory", encoding: "binary", mime_type: "inode/directory"}
   """
-  def perform_once(path, options \\ []) do
+  def perform(path, options \\ [], timeout \\ @process_timeout) do
     with {:ok, pid} <- Server.start_link(options),
-         {:ok, result} <- Server.perform(pid, path),
+         {:ok, result} <- Server.perform(pid, path, timeout),
          :ok <- Server.stop(pid) do
       {:ok, result}
     end