+++ test/majic/helpers_test.exs
@@ -1,9 +1,14 @@
-defmodule Majic.HelpersTest do
+defmodule Majic.OnceTest do
   use Majic.MagicCase
-  doctest Majic.Helpers
+  doctest Majic.Once
 
-  test "perform_once" do
+  test "perform" do
     path = absolute_path("Makefile")
-    assert {:ok, %{mime_type: "text/x-makefile"}} = Majic.Helpers.perform_once(path)
+    assert {:ok, %{mime_type: "text/x-makefile"}} = Majic.Once.perform(path)
+  end
+
+  test "Majic.perform" do
+    path = absolute_path("Makefile")
+    assert {:ok, %{mime_type: "text/x-makefile"}} = Majic.perform(path, once: true)
   end
 end